+++ hadoop-hdfs-project/hadoop-hdfs-rbf/src/test/java/org/apache/hadoop/hdfs/server/federation/router/async/RouterAsyncProtocolTestBase.java
@@ -15,41 +15,37 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.hdfs.server.federation.router;
+package org.apache.hadoop.hdfs.server.federation.router.async;
 
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.FSDataOutputStream;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.fs.QuotaUsage;
-import org.apache.hadoop.fs.StorageType;
 import org.apache.hadoop.fs.permission.FsPermission;
 import org.apache.hadoop.hdfs.server.federation.MiniRouterDFSCluster;
 import org.apache.hadoop.hdfs.server.federation.MockResolver;
 import org.apache.hadoop.hdfs.server.federation.RouterConfigBuilder;
-import org.apache.hadoop.hdfs.server.federation.router.async.AsyncQuota;
-import org.apache.hadoop.hdfs.server.federation.router.async.RouterAsyncRpcClient;
+import org.apache.hadoop.hdfs.server.federation.router.RBFConfigKeys;
+import org.apache.hadoop.hdfs.server.federation.router.RouterRpcServer;
 import org.apache.hadoop.ipc.CallerContext;
 import org.junit.After;
 import org.junit.AfterClass;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.BeforeClass;
-import org.junit.Test;
 import org.mockito.Mockito;
 
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
 
-import static org.apache.hadoop.hdfs.DFSConfigKeys.DFS_QUOTA_BY_STORAGETYPE_ENABLED_KEY;
 import static org.apache.hadoop.hdfs.server.federation.FederationTestUtils.NAMENODES;
 import static org.apache.hadoop.hdfs.server.federation.MiniRouterDFSCluster.DEFAULT_HEARTBEAT_INTERVAL_MS;
 import static org.apache.hadoop.hdfs.server.federation.router.RBFConfigKeys.DFS_ROUTER_RPC_ASYNC_HANDLER_COUNT;
 import static org.apache.hadoop.hdfs.server.federation.router.RBFConfigKeys.DFS_ROUTER_RPC_ASYNC_RESPONDER_COUNT;
-import static org.apache.hadoop.hdfs.server.federation.router.async.utils.AsyncUtil.syncReturn;
 import static org.junit.Assert.assertTrue;
 
-public class TestRouterAsyncQuota {
+/**
+ * Used to test the functionality of async router rps.
+ */
+public class RouterAsyncProtocolTestBase {
   private static Configuration routerConf;
   /** Federated HDFS cluster. */
   private static MiniRouterDFSCluster cluster;
@@ -59,17 +55,14 @@ public class TestRouterAsyncQuota {
   private MiniRouterDFSCluster.RouterContext router;
   private FileSystem routerFs;
   private RouterRpcServer routerRpcServer;
-  private AsyncQuota asyncQuota;
-
-  private final String testfilePath = "/testdir/testAsyncQuota.file";
+  private RouterRpcServer routerAsyncRpcServer;
 
   @BeforeClass
   public static void setUpCluster() throws Exception {
     cluster = new MiniRouterDFSCluster(true, 1, 2,
         DEFAULT_HEARTBEAT_INTERVAL_MS, 1000);
     cluster.setNumDatanodesPerNameservice(3);
-    cluster.setRacks(
-        new String[] {"/rack1", "/rack2", "/rack3"});
+
     cluster.startCluster();
 
     // Making one Namenode active per nameservice
@@ -82,7 +75,6 @@ public class TestRouterAsyncQuota {
     // Start routers with only an RPC service
     routerConf = new RouterConfigBuilder()
         .rpc()
-        .quota(true)
         .build();
 
     // Reduce the number of RPC clients threads to overload the Router easy
@@ -92,7 +84,6 @@ public class TestRouterAsyncQuota {
     // We decrease the DN cache times to make the test faster
     routerConf.setTimeDuration(
         RBFConfigKeys.DN_REPORT_CACHE_EXPIRE, 1, TimeUnit.SECONDS);
-    routerConf.setBoolean(DFS_QUOTA_BY_STORAGETYPE_ENABLED_KEY, true);
     cluster.addRouterOverrides(routerConf);
     // Start routers with only an RPC service
     cluster.startRouters();
@@ -121,19 +112,14 @@ public class TestRouterAsyncQuota {
         routerConf, router.getRouter(), routerRpcServer.getNamenodeResolver(),
         routerRpcServer.getRPCMonitor(),
         routerRpcServer.getRouterStateIdContext());
-    RouterRpcServer spy = Mockito.spy(routerRpcServer);
-    Mockito.when(spy.getRPCClient()).thenReturn(asyncRpcClient);
-    asyncQuota = new AsyncQuota(router.getRouter(), spy);
+    routerAsyncRpcServer = Mockito.spy(routerRpcServer);
+    Mockito.when(routerAsyncRpcServer.getRPCClient()).thenReturn(asyncRpcClient);
 
     // Create mock locations
     MockResolver resolver = (MockResolver) router.getRouter().getSubclusterResolver();
     resolver.addLocation("/", ns0, "/");
     FsPermission permission = new FsPermission("705");
     routerFs.mkdirs(new Path("/testdir"), permission);
-    FSDataOutputStream fsDataOutputStream = routerFs.create(
-        new Path(testfilePath), true);
-    fsDataOutputStream.write(new byte[1024]);
-    fsDataOutputStream.close();
   }
 
   @After
@@ -147,22 +133,31 @@ public class TestRouterAsyncQuota {
     }
   }
 
-  @Test
-  public void testRouterAsyncGetQuotaUsage() throws Exception {
-    asyncQuota.getQuotaUsage("/testdir");
-    QuotaUsage quotaUsage = syncReturn(QuotaUsage.class);
-    // 3-replication.
-    Assert.assertEquals(3 * 1024, quotaUsage.getSpaceConsumed());
-    // We have one directory and one file.
-    Assert.assertEquals(2, quotaUsage.getFileAndDirectoryCount());
+  public static Configuration getRouterConf() {
+    return routerConf;
   }
 
-  @Test
-  public void testRouterAsyncSetQuotaUsage() throws Exception {
-    asyncQuota.setQuota("/testdir", Long.MAX_VALUE, 8096, StorageType.DISK, false);
-    syncReturn(void.class);
-    asyncQuota.getQuotaUsage("/testdir");
-    QuotaUsage quotaUsage = syncReturn(QuotaUsage.class);
-    Assert.assertEquals(8096, quotaUsage.getTypeQuota(StorageType.DISK));
+  public static MiniRouterDFSCluster getCluster() {
+    return cluster;
+  }
+
+  public static String getNs0() {
+    return ns0;
+  }
+
+  public MiniRouterDFSCluster.RouterContext getRouter() {
+    return router;
+  }
+
+  public FileSystem getRouterFs() {
+    return routerFs;
+  }
+
+  public RouterRpcServer getRouterRpcServer() {
+    return routerRpcServer;
+  }
+
+  public RouterRpcServer getRouterAsyncRpcServer() {
+    return routerAsyncRpcServer;
   }
 }
