+++ hadoop-hdfs-project/hadoop-hdfs-rbf/src/test/java/org/apache/hadoop/hdfs/server/federation/router/async/TestRouterAsyncStoragePolicy.java
@@ -15,21 +15,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.hdfs.server.federation.router;
+package org.apache.hadoop.hdfs.server.federation.router.async;
 
 import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.fs.FSDataOutputStream;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.fs.permission.FsPermission;
 import org.apache.hadoop.hdfs.protocol.BlockStoragePolicy;
-import org.apache.hadoop.hdfs.protocol.DatanodeInfo;
-import org.apache.hadoop.hdfs.protocol.HdfsConstants;
 import org.apache.hadoop.hdfs.server.federation.MiniRouterDFSCluster;
 import org.apache.hadoop.hdfs.server.federation.MockResolver;
 import org.apache.hadoop.hdfs.server.federation.RouterConfigBuilder;
-import org.apache.hadoop.hdfs.server.federation.resolver.RemoteLocation;
-import org.apache.hadoop.hdfs.server.federation.router.async.RouterAsyncRpcClient;
-import org.apache.hadoop.hdfs.server.protocol.DatanodeStorageReport;
+import org.apache.hadoop.hdfs.server.federation.router.RBFConfigKeys;
+import org.apache.hadoop.hdfs.server.federation.router.RouterRpcServer;
 import org.apache.hadoop.ipc.CallerContext;
 import org.junit.After;
 import org.junit.AfterClass;
@@ -39,8 +37,6 @@ import org.junit.Test;
 import org.mockito.Mockito;
 
 import java.io.IOException;
-import java.util.List;
-import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
 import static org.apache.hadoop.hdfs.server.federation.FederationTestUtils.NAMENODES;
@@ -48,14 +44,12 @@ import static org.apache.hadoop.hdfs.server.federation.MiniRouterDFSCluster.DEFA
 import static org.apache.hadoop.hdfs.server.federation.router.RBFConfigKeys.DFS_ROUTER_RPC_ASYNC_HANDLER_COUNT;
 import static org.apache.hadoop.hdfs.server.federation.router.RBFConfigKeys.DFS_ROUTER_RPC_ASYNC_RESPONDER_COUNT;
 import static org.apache.hadoop.hdfs.server.federation.router.async.utils.AsyncUtil.syncReturn;
+import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertTrue;
 
-/**
- * Used to test the async functionality of {@link RouterRpcServer}.
- */
-public class TestRouterAsyncRpcServer {
+public class TestRouterAsyncStoragePolicy {
   private static Configuration routerConf;
   /** Federated HDFS cluster. */
   private static MiniRouterDFSCluster cluster;
@@ -64,7 +58,10 @@ public class TestRouterAsyncRpcServer {
   /** Random Router for this federated cluster. */
   private MiniRouterDFSCluster.RouterContext router;
   private FileSystem routerFs;
-  private RouterRpcServer asyncRouterRpcServer;
+  private RouterRpcServer routerRpcServer;
+  private RouterAsyncStoragePolicy asyncStoragePolicy;
+
+  private final String testfilePath = "/testdir/testAsyncStoragePolicy.file";
 
   @BeforeClass
   public static void setUpCluster() throws Exception {
@@ -115,20 +112,25 @@ public class TestRouterAsyncRpcServer {
   public void setUp() throws IOException {
     router = cluster.getRandomRouter();
     routerFs = router.getFileSystem();
-    RouterRpcServer routerRpcServer = router.getRouterRpcServer();
+    routerRpcServer = router.getRouterRpcServer();
     routerRpcServer.initAsyncThreadPool();
     RouterAsyncRpcClient asyncRpcClient = new RouterAsyncRpcClient(
         routerConf, router.getRouter(), routerRpcServer.getNamenodeResolver(),
         routerRpcServer.getRPCMonitor(),
         routerRpcServer.getRouterStateIdContext());
-    asyncRouterRpcServer = Mockito.spy(routerRpcServer);
-    Mockito.when(asyncRouterRpcServer.getRPCClient()).thenReturn(asyncRpcClient);
+    RouterRpcServer spy = Mockito.spy(routerRpcServer);
+    Mockito.when(spy.getRPCClient()).thenReturn(asyncRpcClient);
+    asyncStoragePolicy = new RouterAsyncStoragePolicy(spy);
 
     // Create mock locations
     MockResolver resolver = (MockResolver) router.getRouter().getSubclusterResolver();
     resolver.addLocation("/", ns0, "/");
     FsPermission permission = new FsPermission("705");
     routerFs.mkdirs(new Path("/testdir"), permission);
+    FSDataOutputStream fsDataOutputStream = routerFs.create(
+        new Path(testfilePath), true);
+    fsDataOutputStream.write(new byte[1024]);
+    fsDataOutputStream.close();
   }
 
   @After
@@ -142,51 +144,22 @@ public class TestRouterAsyncRpcServer {
     }
   }
 
-  /**
-   * Test that the async RPC server can invoke a method at an available Namenode.
-   */
   @Test
-  public void testInvokeAtAvailableNsAsync() throws Exception {
-    RemoteMethod method = new RemoteMethod("getStoragePolicies");
-    asyncRouterRpcServer.invokeAtAvailableNsAsync(method, BlockStoragePolicy[].class);
-    BlockStoragePolicy[] storagePolicies = syncReturn(BlockStoragePolicy[].class);
-    assertEquals(8, storagePolicies.length);
-  }
+  public void testRouterAsyncStoragePolicy() throws Exception {
+    BlockStoragePolicy[] storagePolicies = cluster.getNamenodes().get(0)
+        .getClient().getStoragePolicies();
+    asyncStoragePolicy.getStoragePolicies();
+    BlockStoragePolicy[] storagePoliciesAsync = syncReturn(BlockStoragePolicy[].class);
+    assertArrayEquals(storagePolicies, storagePoliciesAsync);
 
-  /**
-   * Test get create location async.
-   */
-  @Test
-  public void testGetCreateLocationAsync() throws Exception {
-    final List<RemoteLocation> locations =
-        asyncRouterRpcServer.getLocationsForPath("/testdir", true);
-    asyncRouterRpcServer.getCreateLocationAsync("/testdir", locations);
-    RemoteLocation remoteLocation = syncReturn(RemoteLocation.class);
-    assertNotNull(remoteLocation);
-    assertEquals(ns0, remoteLocation.getNameserviceId());
-  }
+    asyncStoragePolicy.getStoragePolicy(testfilePath);
+    BlockStoragePolicy blockStoragePolicy1 = syncReturn(BlockStoragePolicy.class);
 
-  /**
-   * Test get datanode report async.
-   */
-  @Test
-  public void testGetDatanodeReportAsync() throws Exception {
-    asyncRouterRpcServer.getDatanodeReportAsync(
-        HdfsConstants.DatanodeReportType.ALL, true, 0);
-    DatanodeInfo[] datanodeInfos = syncReturn(DatanodeInfo[].class);
-    assertEquals(3, datanodeInfos.length);
-
-    // Get the namespace where the datanode is located
-    asyncRouterRpcServer.getDatanodeStorageReportMapAsync(HdfsConstants.DatanodeReportType.ALL);
-    Map<String, DatanodeStorageReport[]> map = syncReturn(Map.class);
-    assertEquals(1, map.size());
-    assertEquals(3, map.get(ns0).length);
-
-    DatanodeInfo[] slowDatanodeReport1 =
-        asyncRouterRpcServer.getSlowDatanodeReport(true, 0);
-
-    asyncRouterRpcServer.getSlowDatanodeReportAsync(true, 0);
-    DatanodeInfo[] slowDatanodeReport2 = syncReturn(DatanodeInfo[].class);
-    assertEquals(slowDatanodeReport1, slowDatanodeReport2);
+    asyncStoragePolicy.setStoragePolicy(testfilePath, "COLD");
+    syncReturn(null);
+    asyncStoragePolicy.getStoragePolicy(testfilePath);
+    BlockStoragePolicy blockStoragePolicy2 = syncReturn(BlockStoragePolicy.class);
+    assertNotEquals(blockStoragePolicy1, blockStoragePolicy2);
+    assertEquals("COLD", blockStoragePolicy2.getName());
   }
 }
